Add Figure 7: Posterior median associated
Figure 7: Posterior median associated with the African buffalo data analysis for the a) recharge function g(t) shown as color on top of the posterior mean trajectory μ(t). Distance to surface water is shown in the background for reference (with small distances indicated by darker shades) and green indicating inferred areas associated with recharge. Map in (a) oriented such that north is up. Posterior median trajectories (b, c) and d) recharge function g(t) and e) decision probability ρ(t) with 95% credible intervals shown in gray with posterior mean for the decision z(t) shown as black points. Color corresponds to the value of the recharge function. Distance to surface water profile shown as gray line in (b) and (c) for reference. Green rug at the bottom of (b) represents times when the recharge function is increasing.