Add Figure 8: a) Simulated trajectory μ(t)
Figure 8: a) Simulated trajectory μ(t) in two dimensions and b) associated recharge (g(t)) and c) decision probability (ρ(t)). The green polygon shown in the middle of (a) represents a patch that corresponds to recharge and the single covariate based on distance to patch is shown in (a) as shading from light (near patch) to dark (far from patch). The trajectory color corresponds to the time elapsed since the initial point of the study period (increasing from dark blue to yellow). The simulated decision process z(t) is shown as semitransparent points at the top (z(t)=1) and bottom (z(t)=0) of the plot for ρ(t) in (c).