Add Figure 10: Telemetry data (points) and posterior mean trajectory
Figure 10: Telemetry data (points) and posterior mean trajectory (multicolored line) for the a) easting and b) northing directions. Gray lines in (a) and (b) represent the value of x(μ(t)) and green rug plot below (a) represents positions where the recharge function is increasing. The posterior median c) recharge function g(t) and d) associated probability ρ(t) of decision to recharge are shown as multicolored lines with 95% credible intervals shown in gray (z(t) shown as black points). True values of g(t) and ρ(t) are shown as red lines. The color associated with the estimates indicates discharged when orange and charged when blue.