Deep Neural Networks as Point Estimates for Deep Gaussian Processes
Deep Gaussian processes (DGPs) have struggled for relevance in applications due to the challenges
and cost associated with Bayesian inference. In this paper we propose a sparse variational approximation
for DGPs for which the approximate posterior mean has the same mathematical structure as a Deep
Neural Network (DNN). We make the forward pass through a DGP equivalent to a ReLU DNN by finding
an interdomain transformation that represents the GP posterior mean as a sum of ReLU basis functions.
This unification enables the initialisation and training of the DGP as a neural network, leveraging the
well established practice in the deep learning community, and so greatly aiding the inference task. The
experiments demonstrate improved accuracy and faster training compared to current DGP methods,
while retaining favourable predictive uncertainties.