SYNTHETIC DATA FOR MODEL SELECTION
Recent improvements in synthetic data generation make it possible to produce im-
ages that are highly photorealistic and indistinguishable from real ones. Further-
more, synthetic generation pipelines have the potential to generate an unlimited
number of images. The combination of high photorealism and scale turn the syn-
thetic data into a promising candidate for potentially improving various machine
learning (ML) pipelines. Thus far, a large body of research in this field has focused
on using synthetic images for training, by augmenting and enlarging training data.
In contrast to using synthetic data for training, in this work we explore whether
synthetic data can be beneficial for model selection. Considering the task of image
classification, we demonstrate that when data is scarce, synthetic data can be used
to replace the held out validation set, thus allowing to train on a larger dataset.