DeepSMOTE: Fusing Deep Learning and SMOTE for Imbalanced Data
Abstract—Despite over two decades of progress, imbalanced
data is still considered a significant challenge for contemporary
machine learning models. Modern advances in deep learning
have magnified the importance of the imbalanced data problem.
The two main approaches to address this issue are based on
loss function modifications and instance resampling. Instance
sampling is typically based on Generative Adversarial Networks
(GANs), which may suffer from mode collapse. Therefore, there
is a need for an oversampling method that is specifically
tailored to deep learning models, can work on raw images
while preserving their properties, and is capable of generating
high quality, artificial images that can enhance minority classes
and balance the training set. We propose DeepSMOTE - a
novel oversampling algorithm for deep learning models. It is
simple, yet effective in its design. It consists of three major
components: (i) an encoder/decoder framework; (ii) SMOTE-
based oversampling; and (iii) a dedicated loss function that
is enhanced with a penalty term. An important advantage of
DeepSMOTE over GAN-based oversampling is that DeepSMOTE
does not require a discriminator, and it generates high-quality
artificial images that are both information-rich and suitable for
visual inspection. DeepSMOTE code is publicly available at:
https://github.com/dd1github/DeepSMOTE.