PyTorch Tabular: A Framework for Deep Learning with Tabular Data
Abstract — In spite of showing unreasonable effectiveness 
in modalities like Text and Image, Deep Learning has always 
lagged Gradient Boosting in tabular data—both in popularity
and performance. But recently there have been newer models 
created specifically for tabular data, which is pushing the 
performance bar. But popularity is still a challenge because 
there is no easy, ready-to-use library like Sci-Kit Learn for 
deep learning. PyTorch Tabular is a new deep learning library 
which makes working with Deep Learning and tabular data 
easy and fast. It is a library built on top of PyTorch and 
PyTorch Lightning and works on pandas dataframes directly. 
Many SOTA models like NODE and TabNet are already 
integrated and implemented in the library with a unified API.
PyTorch Tabular is designed to be easily extensible for 
researchers, simple for practitioners, and robust in industrial 
deployments. The library is available at 
https://github.com/manujosephv/pytorch_tabular